+++ scyan/plot/ratios.py
@@ -7,6 +7,7 @@ import numpy as np
 import seaborn as sns
 from anndata import AnnData
 from matplotlib.lines import Line2D
+from scipy.cluster import hierarchy
 
 from ..tools import cell_type_ratios
 from .utils import plot_decorator
@@ -20,6 +21,7 @@ def pop_percentage(
     groupby: Union[str, List[str], None] = None,
     key: str = "scyan_pop",
     figsize: tuple[float, float] = None,
+    dendogram: bool = False,
     show: bool = True,
 ):
     """Show populations percentages. Depending on `groupby`, this is either done globally, or as a stacked bar plot (one bar for each group).
@@ -29,14 +31,21 @@ def pop_percentage(
         groupby: Key(s) of `adata.obs` used to create groups (e.g. the patient ID).
         key: Key of `adata.obs` containing the population names (or the values) for which percentage will be displayed.
         figsize: matplotlib figure size.
+        dendogram: If True, the groups are sorted based on a dendogram clustering.
         show: Whether or not to display the figure.
     """
     if groupby is None:
         adata.obs[key].value_counts(normalize=True).mul(100).plot.bar(figsize=figsize)
     else:
-        adata.obs.groupby(groupby)[key].value_counts(normalize=True).mul(
-            100
-        ).unstack().plot.bar(stacked=True, figsize=figsize)
+        df = adata.obs.groupby(groupby)[key].value_counts(normalize=True)
+        df = df.mul(100).unstack()
+
+        if dendogram:
+            Z = hierarchy.linkage(df.values, method="ward")
+            dendrogram = hierarchy.dendrogram(Z, no_plot=True)
+            df = df.iloc[dendrogram["leaves"]]
+
+        df.plot.bar(stacked=True, figsize=figsize)
         plt.legend(
             bbox_to_anchor=(1.04, 0.5), loc="center left", borderaxespad=0, frameon=False
         )
@@ -81,7 +90,7 @@ def pop_dynamics(
         groupby = ([groupby] if isinstance(groupby, str) else groupby) + [time_key]
 
     df = cell_type_ratios(adata, groupby=groupby, key=key, normalize="%", among=among)
-    df.index = df.index.set_levels(df.index.levels[-1].astype(str), level=-1)
+    df.index = df.index.set_levels(df.index.levels[-1].codes, level=-1)
     df_log_count = np.log(
         1 + cell_type_ratios(adata, groupby=groupby, key=key, normalize=False)
     )